Text version is working
Only one slight error:
If the player attempts an illegal move such as moving left or right out of the bounds of the board, then the move should essentially be "skipped". The movers affected will therefore initially move when they shouldn't, although they will correct themselves in one turn. The effect is therefore only graphical (as far as I'm aware -- need to check this properly).